Extract Remediation class for point calculation
