+++ lib/cc/engine/bundler_audit.rb
@@ -4,6 +4,7 @@ require "json"
 require "versionomy"
 
 require "cc/engine/bundler_audit/analyzer"
+require "cc/engine/bundler_audit/remediation"
 require "cc/engine/bundler_audit/result"
 
 module CC
